topic-creator/requirements.txt
chaining==0.0.2
kafka-python==2.0.2
pyChaining==0.0.2
pyXhUtils==0.0.12
